Reduce risk of overflow via common denominators
When implementing `r1 + r2`, check whether they have the same
denominator, and if so skip the generic multiplication-of-denominators.
This is not as comprehensive as using `gcd`, but it is *much* faster,
and that is a key design goal for this package.

For consumers, one can exploit this by ensuring that all coefficients
for a single axis have the same denominator.  For example, if you
wanted to construct a weighted average of `a`, `b`, and `c`, rather
than

    cedge, cmid = SimpleRatio(1, 4), SimpleRatio(1, 2)
    cedge*a + cmid*b + cedge*c

it would be better to use

    cedge, cmid = SimpleRatio(1, 4), SimpleRatio(2, 4)

Fixes https://github.com/JuliaMath/Interpolations.jl/issues/457